requirements.txt
aiogram==3.10.0
google-auth==2.23.0
gspread==5.11.3
python-dotenv==1.0.0
uvicorn==0.33.0